+++ examples/example-jetstream-deadletter/src/test/java/org/example/natspring/jetstreamdeadletter/JetStreamDeadLetterExampleTests.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 
 import io.github.amadeusitgroup.testcontainers.nats.NatsContainer;
-import io.github.malczuuu.natspring.core.NatsClient;
+import io.github.malczuuu.natspring.core.NatsOperations;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import org.junit.jupiter.api.BeforeEach;
@@ -24,7 +24,7 @@ class JetStreamDeadLetterExampleTests {
   public static final NatsContainer nats = new NatsContainer("nats:2.14").withJetStream();
 
   @Autowired JetStreamDeadLetterExample application;
-  @Autowired NatsClient natsClient;
+  @Autowired NatsOperations natsClient;
   @Autowired RestTestClient restClient;
 
   @BeforeEach