+++ integration-tests/src/test/java/io/github/malczuuu/natspring/itest/GsonConverterTests.java
@@ -19,7 +19,7 @@ package io.github.malczuuu.natspring.itest;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.github.malczuuu.natspring.converter.gson.GsonNatsMessageConverter;
-import io.github.malczuuu.natspring.core.NatsClient;
+import io.github.malczuuu.natspring.core.NatsOperations;
 import io.github.malczuuu.natspring.itest.entrypoint.Entrypoint;
 import io.github.malczuuu.natspring.itest.entrypoint.NatsListenerComponent;
 import io.github.malczuuu.natspring.itest.entrypoint.SampleMessage;
@@ -46,7 +46,7 @@ class GsonConverterTests extends AbstractSpringBootTests {
   }
 
   @Autowired private NatsListenerComponent handler;
-  @Autowired private NatsClient natsClient;
+  @Autowired private NatsOperations natsClient;
 
   @AfterEach
   void afterEach() {