+++ integration-tests/src/test/java/io/github/malczuuu/natspring/itest/NatsListenerTests.java
@@ -18,7 +18,7 @@ package io.github.malczuuu.natspring.itest;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.github.malczuuu.natspring.core.NatsClient;
+import io.github.malczuuu.natspring.core.NatsOperations;
 import io.github.malczuuu.natspring.itest.entrypoint.NatsListenerComponent;
 import io.github.malczuuu.natspring.itest.entrypoint.SampleMessage;
 import io.github.malczuuu.natspring.itest.fixture.AbstractSpringBootTests;
@@ -35,7 +35,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 class NatsListenerTests extends AbstractSpringBootTests {
 
   @Autowired private NatsListenerComponent handler;
-  @Autowired private NatsClient natsClient;
+  @Autowired private NatsOperations natsClient;
 
   @AfterEach
   void afterEach() {